feat(adieu-razzle): moved all build scripts into single file * also added size script * also configured WebpackBar for each kind of build
