+++ examples/shop-with-blog/client/package.json
@@ -10,7 +10,7 @@
     "start:dbg-brk": "cross-env falcon-client start -- --inspect-brk",
     "start:prod": "cross-env NODE_ENV=production node build/server.js",
     "build": "falcon-client build",
-    "size": "cross-env NODE_ANALYZE=true npm run build",
+    "size": "falcon-client size",
     "test": "falcon-client test --env=jsdom",
     "test:coverage": "falcon-client test --env=jsdom --ci --coverage",
     "lint": "eslint ./"
