+++ packages/falcon-client/bin/falcon-client.js
@@ -7,8 +7,10 @@ process.on('uncaughtException', ex => {
   console.log('Uncaught Exception: ', ex);
 });
 
+process.noDeprecation = true; // turns off that loadQuery clutter.
+
 const Logger = require('@deity/falcon-logger');
-const webpack = require('./../build-utils/webpack');
+const { startDevServer, build, size } = require('./../build-utils/webpack');
 const workbox = require('./../src/buildTools/workbox');
 const jest = require('./../build-utils/jest');
 
@@ -19,18 +21,25 @@ const jest = require('./../build-utils/jest');
   try {
     switch (script) {
       case 'start': {
-        await webpack.startDevServer();
+        await startDevServer();
         break;
       }
+
       case 'build': {
-        await webpack.build();
+        await build();
         await workbox.injectManifest();
         break;
       }
+      case 'size': {
+        await size();
+        break;
+      }
+
       case 'test': {
         jest();
         break;
       }
+
       default:
         Logger.log(`Unknown script "${script}".`);
         Logger.log('Perhaps you need to update @deity/falcon-client?');
