+++ packages/falcon-client/build-utils/webpack/index.js
@@ -1,4 +1,178 @@
-module.exports = {
-  startDevServer: require('./start'),
-  build: require('./build')
+const fs = require('fs-extra');
+const chalk = require('chalk');
+const Logger = require('@deity/falcon-logger');
+const WebpackDevServer = require('webpack-dev-server-speedy');
+const clearConsole = require('react-dev-utils/clearConsole');
+const checkRequiredFiles = require('react-dev-utils/checkRequiredFiles');
+const { choosePort } = require('react-dev-utils/WebpackDevServerUtils');
+const { measureFileSizesBeforeBuild, printFileSizesAfterBuild } = require('react-dev-utils/FileSizeReporter');
+
+const paths = require('./../paths');
+const { getBuildConfig, removePreviousBuildAssets, webpackCompiler, webpackCompileAsync } = require('./tools');
+const createConfig = require('./config/create');
+
+module.exports.startDevServer = async () => {
+  if (!checkRequiredFiles([paths.appIndexJs])) {
+    process.exit(1);
+  }
+
+  process.env.NODE_ENV = process.env.NODE_ENV || 'development';
+  process.env.BABEL_ENV = process.env.NODE_ENV;
+  process.env.HOST = process.env.HOST || 'localhost';
+  process.env.PORT = await choosePort(process.env.HOST, parseInt(process.env.PORT, 10) || 3000);
+
+  try {
+    const falconConfig = getBuildConfig();
+    if (falconConfig.clearConsole) {
+      clearConsole();
+    }
+
+    Logger.info(chalk`{hex('#a9cf38') Compiling...}`);
+    removePreviousBuildAssets(paths.appBuild, paths.appBuildPublic);
+
+    const options = {
+      env: process.env.NODE_ENV,
+      host: process.env.HOST,
+      port: parseInt(process.env.PORT, 10),
+      devServerPort: parseInt(process.env.PORT, 10) + 1,
+      inspect: process.argv.find(x => x.match(/--inspect-brk(=|$)/) || x.match(/--inspect(=|$)/)) || undefined
+    };
+
+    const clientConfig = createConfig('web', options, falconConfig);
+    const serverConfig = createConfig('node', options, falconConfig);
+
+    // Compile our assets with webpack
+    const clientCompiler = webpackCompiler(clientConfig);
+    const serverCompiler = webpackCompiler(serverConfig);
+
+    // Start our server webpack instance in watch mode after assets compile
+    clientCompiler.plugin('done', () => {
+      serverCompiler.watch(
+        {
+          quiet: true,
+          stats: 'none'
+        },
+        /* eslint-disable no-unused-vars */
+        stats => {}
+      );
+    });
+
+    // Create a new instance of Webpack-dev-server for our client assets.
+    const clientDevServer = new WebpackDevServer(clientCompiler, clientConfig.devServer);
+    clientDevServer.listen(options.devServerPort, error => {
+      if (error) {
+        Logger.error(error);
+      }
+    });
+  } catch (error) {
+    Logger.error('Compilation failed!');
+
+    throw error;
+  }
+};
+
+module.exports.build = async () => {
+  if (!checkRequiredFiles([paths.appIndexJs])) {
+    process.exit(1);
+  }
+
+  process.env.NODE_ENV = 'production';
+  process.env.BABEL_ENV = process.env.NODE_ENV;
+
+  try {
+    const falconConfig = getBuildConfig();
+    if (falconConfig.clearConsole) {
+      clearConsole();
+    }
+
+    const options = {
+      env: process.env.NODE_ENV,
+      publicPath: process.env.PUBLIC_PATH || '/',
+      isCI: process.env.CI && (typeof process.env.CI !== 'string' || process.env.CI.toLowerCase() !== 'false')
+    };
+
+    Logger.log('Creating an optimized production build...');
+    const previousBuildSizes = await measureFileSizesBeforeBuild(paths.appBuildPublic);
+    fs.emptyDirSync(paths.appBuild);
+    fs.copySync(paths.appPublic, paths.appBuildPublic, { dereference: true });
+
+    // First compile the client. We need it to properly output assets.json
+    // (asset manifest file with the correct hashes on file names BEFORE we can start the server compiler).
+
+    const clientConfig = createConfig('web', options, falconConfig);
+    const clientCompilation = await webpackCompileAsync(clientConfig, options.isCI);
+
+    const serverConfig = createConfig('node', options, falconConfig);
+    // ContextReplacementPlugin https://webpack.js.org/plugins/context-replacement-plugin/
+    /* const serverCompilation = */ await webpackCompileAsync(serverConfig, options.isCI);
+
+    const warnings = [...clientCompilation.warnings]; // , ...serverCompilation.warnings]
+
+    if (warnings.length) {
+      Logger.warn(chalk.yellow('\nCompiled with warnings.\n'));
+      Logger.warn(warnings.join('\n\n'));
+      Logger.log();
+    } else {
+      Logger.log(chalk.green('\nCompiled successfully.\n'));
+    }
+
+    Logger.log('File sizes after gzip:\n');
+    const { stats } = clientCompilation;
+    printFileSizesAfterBuild(stats, previousBuildSizes, paths.appBuild);
+
+    Logger.log();
+  } catch (error) {
+    Logger.error(`${chalk.red('\nFailed to compile.\n')}`);
+    Logger.error(error);
+    Logger.log();
+
+    process.exit(1);
+  }
+};
+
+module.exports.size = async () => {
+  if (!checkRequiredFiles([paths.appIndexJs])) {
+    process.exit(1);
+  }
+
+  process.env.NODE_ENV = 'production';
+  process.env.BABEL_ENV = process.env.NODE_ENV;
+
+  try {
+    const falconConfig = getBuildConfig();
+    if (falconConfig.clearConsole) {
+      clearConsole();
+    }
+
+    const options = {
+      env: process.env.NODE_ENV,
+      publicPath: process.env.PUBLIC_PATH || '/',
+      analyze: true
+    };
+
+    Logger.log('Creating an optimized production build...');
+
+    fs.emptyDirSync(paths.appBuild);
+    fs.copySync(paths.appPublic, paths.appBuildPublic, { dereference: true });
+
+    Logger.log('Compiling client...');
+    const clientConfig = createConfig('web', options, falconConfig);
+    const { warnings } = await webpackCompileAsync(clientConfig);
+
+    if (warnings.length) {
+      Logger.warn(chalk.yellow('\nCompiled client with warnings.\n'));
+      Logger.warn(warnings.join('\n\n'));
+      Logger.log();
+    } else {
+      Logger.log(chalk.green('\nCompiled client successfully.\n'));
+    }
+
+    Logger.log();
+  } catch (error) {
+    Logger.error(`${chalk.red('\nFailed to compile client.\n')}`);
+    Logger.error(error);
+    Logger.log();
+
+    process.exit(1);
+  }
 };
